+++ core/apis_tps.go
@@ -33,6 +33,11 @@ func (a appTPS) CreateEvents(event []model.LegacyEventItem) ([]model.LegacyEvent
 	return a.app.storage.InsertLegacyEvents(nil, event)
 }
 
+// DeleteEvents deletes legacy events by ids and creator
+func (a appTPS) DeleteEvents(ids []string, accountID string) error {
+	return a.app.storage.DeleteLegacyEventsByIDsAndCreator(nil, ids, accountID)
+}
+
 // newAppTPS creates new appTPS
 func newAppTPS(app *Application) appTPS {
 	return appTPS{app: app}